Move unit tests for HuggingFaceHub util fns to own file
